Three leg vertices version!
* Renamed a lot of variables to better reflect their content
* Implemented gamma^w, gamma^q, and eta^q
* Restructured the program a bit to avoid too many intermediate quantities
* The do_chi part will currently only calculate chi^q - chi^w, since chi^w should really be taken directly from the ctqmc solver and thuis read from file.
* The code compiles! But I have not tested it! I might have fixed a bug in gamma^w, but I might also have introduced some new bugs..

+++ vq_module.f90
@@ -16,7 +16,7 @@ subroutine read_vq(iq, v)
   complex(kind=8) :: vq(ndim,ndim,ndim,ndim)
   integer(hid_t) :: vq_file_id, grp_id, iq_id, iq_space_id
   integer :: err, ind, i, j, k, l, i1, i2
-  integer(hid_t) :: nmembers, imembers, itype
+  integer :: nmembers, imembers, itype
   character(len=20) :: name_buffer
   integer(hsize_t), dimension(2) :: iq_dims, iq_maxdims
   integer(hsize_t), dimension(1) :: vq_dims
